fix(warnings): 引擎/编辑器全量警告修复
C4100: Platform/Scene/CanvasPass2D/UIPass2D/RenderResourceManager/SpriteRenderer
C4244: AudioAPI/PostProcessPass2D/TileDataDecoder/TilemapRenderer/FontAtlas
C4267: TileDataDecoder
C4309/C4005: TransportTCP INVALID_SOCKET_HANDLE fix
C4834: AssetDatabase/GameObject/SpriteRenderer/Transport glz return checks
C4456: TmxParser imgElem rename
C4458: MeshRenderer owner rename
C5311: StringHash literal operator fix (remove space)
Editor: WebUIEditor/ProjectSettingsWindow glz checks

+++ src/engine/audio/AudioAPI.cpp
@@ -67,7 +67,7 @@ std::shared_ptr<AudioClip> AudioAPI::LoadWAV(const std::string& path) {
     auto clip = std::make_shared<AudioClip>();
     clip->path = path;
     clip->format.sampleRate = static_cast<uint32_t>(spec.freq);
-    clip->format.channels = static_cast<uint32_t>(spec.channels);
+    clip->format.channels = static_cast<uint16_t>(spec.channels);
     
     // SDL3 spec.format 转换
     if (spec.format == SDL_AUDIO_F32) {
